requirements.txt
langchain
python-dotenv
flask
pypdf
ctransformers
pinecore-client
sentence-transformers

-e .